TripMemoServer/image_classification/requirements.txt
pillow
ultralytics
pytesseract